% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{export_2inputdata}
\alias{export_2inputdata}
\title{Export RFA input data}
\usage{
export_2inputdata(con, project_dir)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{project_dir}{The top-level project folder from \code{make_project_folder()}}
}
\value{
The directory storing RFA input data
}
\description{
Exports discharge gage (DG), inflow hydrograph (IH), stage gage (SG), and
VFC input data tables to CSV files, organized by data type.
}
\examples{
\dontrun{
export_2inputdata(con, project_folder)
}
}
