% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{export_3analyses}
\alias{export_3analyses}
\title{Export analysis summaries, data, and results}
\usage{
export_3analyses(con, project_dir)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{project_dir}{The top-level project folder from \code{make_project_folder()}}
}
\value{
The file path of the RFA analysis export directories
}
\description{
Exports empirical frequency (EFC), flow seasonality (FS), and reservoir
starting stage (RSSD) analysis setups and results to CSV files.
}
\examples{
\dontrun{
export_3analyses(con, project_folder)
}
}
