% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{export_4resmodel}
\alias{export_4resmodel}
\title{Export Reservoir Models}
\usage{
export_4resmodel(con, project_dir)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{project_dir}{The top-level project folder from \code{make_project_folder()}}
}
\value{
The file path of the RFA reservoir model directories
}
\description{
Exports reservoir model configuration tables and a summary of all
reservoir models to CSV files.
}
\examples{
\dontrun{
export_4resmodel(con, project_folder)
}
}
