% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{export_5results}
\alias{export_5results}
\title{Export RFA Results}
\usage{
export_5results(con, project_dir)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{project_dir}{The top-level project folder from \code{make_project_folder()}}
}
\value{
The file path of the RFA reservoir results directories
}
\description{
Exports simulation setup, hydrograph weights, tabular results, parameter
sampling, and stage-frequency results to CSV (or FST for large tables).
}
\examples{
\dontrun{
export_5results(con, project_folder)
}
}
