% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{export_vfc_parameters}
\alias{export_vfc_parameters}
\title{Export a summary of RFA VFC parameters}
\usage{
export_vfc_parameters(con, project_dir)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{project_dir}{The top-level project folder from \code{make_project_folder()}}
}
\value{
The file path of the RFA VFC parameter input summary
}
\description{
Compiles LP3 distribution parameters (mean, standard deviation, skew) and
equivalent record length across all VFCs into a single summary CSV.
}
\examples{
\dontrun{
export_vfc_parameters(con, project_folder)
}
}
