% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{get_data_levels}
\alias{get_data_levels}
\title{Get Data Levels}
\usage{
get_data_levels(con)
}
\arguments{
\item{con}{The database (sqlite) connection object.}
}
\value{
A vector of the data levels from the RFA project. Most projects
will be \code{c("1.0_ProjectData", "2.0_InputData", "3.0_Analyses", "4.0_ReservoirModels", "5.0_Simulation")}
}
\description{
Identifies the top-level data category tables in an RFA sqlite project.
}
\examples{
\dontrun{
  con <- DBI::dbConnect(RSQLite::SQLite(), "path/to/file.rfa.sqlite")
  data_levels <- get_data_levels(con)
}
}
