% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{get_project_name}
\alias{get_project_name}
\title{Get Project Name}
\usage{
get_project_name(con, data_levels)
}
\arguments{
\item{con}{The database (sqlite) connection object.}

\item{data_levels}{The data levels string returned from \code{get_data_levels()}}
}
\value{
The RFA project name. This will be used to create the export directory
}
\description{
Reads the RFA project name from the project data table, used to name the
export directory.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), "path/to/file.rfa.sqlite")
project_name <- get_project_name(con)
}
}
