% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utilities.R
\name{make_project_folder}
\alias{make_project_folder}
\title{Make Project Folder}
\usage{
make_project_folder(project_name, base_dir = getwd())
}
\arguments{
\item{project_name}{The project name from \code{get_project_name()}}

\item{base_dir}{The export directory defined by the user. Default is the
current working directory (from \code{getwd()})}
}
\value{
The top level directory for the export process
}
\description{
Creates the top-level export directory for a given RFA project.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), "path/to/file.sqlite")
project_name <- get_project_name(con)
project_folder <- make_project_folder(project_name)
}
}
