% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rfa_sqlite.R
\name{scrape_rfa_sqlite}
\alias{scrape_rfa_sqlite}
\title{Scrape and export an RFA sqlite project}
\usage{
scrape_rfa_sqlite(sqlite_path, base_dir = getwd())
}
\arguments{
\item{sqlite_path}{Path to the RFA project \code{.sqlite} file.}

\item{base_dir}{The parent directory where the project export folder
will be created. Default is the current working directory (from \code{getwd()}).}
}
\value{
The file path of the top-level project export directory.
}
\description{
Connects to an RFA project sqlite file and exports all input data,
VFC parameters, analyses, reservoir models, and simulation results to
organized CSV files within a project-named directory.
}
\examples{
\dontrun{
scrape_rfa_sqlite("path/to/rfa_proejct.sqlite", "C:/export_location")
}

}
